Enfore compute mode for xdimsum.cl
This is required because (as other packages) the package definition
here is with a comma:

    package xdimsum, bin = ...

In command mode, parameters of package are delimited by spaces, so the
package name is with a comma.

In command mode, thres one needs to add commas also between the
arguments of "hidetask" and "cache", which is done here as well.